Pulling final version of project

+++ src/main/java/org/kylecodes/gm/services/ExerciseServiceImpl.java
@@ -1,21 +1,20 @@
 package org.kylecodes.gm.services;
 
+import org.kylecodes.gm.constants.RequestFailure;
 import org.kylecodes.gm.dtos.ExerciseDto;
 import org.kylecodes.gm.entities.Exercise;
 import org.kylecodes.gm.entities.User;
 import org.kylecodes.gm.entities.Workout;
 import org.kylecodes.gm.exceptions.ExerciseNotFoundException;
-import org.kylecodes.gm.constants.RequestFailure;
 import org.kylecodes.gm.exceptions.WorkoutNotFoundException;
-import org.kylecodes.gm.utils.SecurityUtil;
-import org.kylecodes.gm.repositories.ExerciseRepository;
-import org.kylecodes.gm.repositories.WorkoutRepository;
 import org.kylecodes.gm.mappers.EntityToDtoMapper;
 import org.kylecodes.gm.mappers.ExerciseToExerciseDtoMapper;
+import org.kylecodes.gm.repositories.ExerciseRepository;
+import org.kylecodes.gm.repositories.WorkoutRepository;
+import org.kylecodes.gm.utils.SecurityUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -26,37 +25,14 @@ public class ExerciseServiceImpl implements ExerciseService{
     @Autowired
     private WorkoutRepository workoutRepository;
 
+
     EntityToDtoMapper<Exercise, ExerciseDto> exerciseMapper = new ExerciseToExerciseDtoMapper();
+    
 
     @Override
-    public List<ExerciseDto> getAllExercises() {
-        List<Exercise> exercises = exerciseRepository.findAll();
-
-        return exercises.stream().map((exercise) -> exerciseMapper.mapToDto(exercise)).toList();
-    }
-
-    @Override
-    public List<ExerciseDto> getAllExercisesInWorkout(Long workoutId) {
+    public ExerciseDto createExerciseInWorkout(ExerciseDto exerciseDto, Long workoutId) {
         User user = SecurityUtil.getPrincipalFromSecurityContext();
-        Optional<Workout> workout = Optional.ofNullable(workoutRepository.findByIdAndUser(workoutId, user)
-                .orElseThrow(() -> new WorkoutNotFoundException(RequestFailure.PUT_REQUEST_FAILURE)));
-        List<Exercise> exercisesInWorkout = exerciseRepository.findAllByWorkout(workout.get());
-        return exercisesInWorkout.stream().map((exercise) -> exerciseMapper.mapToDto(exercise)).toList();
-    }
-
-    @Override
-    public ExerciseDto getExerciseInWorkoutById(Long workoutId, Long exerciseId) {
-        User user = SecurityUtil.getPrincipalFromSecurityContext();
-        Optional<Workout> workout = Optional.ofNullable(workoutRepository.findByIdAndUser(workoutId, user)
-                .orElseThrow(() -> new WorkoutNotFoundException(RequestFailure.PUT_REQUEST_FAILURE)));
-        Exercise exercise = exerciseRepository.findByIdAndWorkout(exerciseId, workout.get()).orElseThrow(() -> new ExerciseNotFoundException(RequestFailure.GET_REQUEST_FAILURE));
-        return exerciseMapper.mapToDto(exercise);
-    }
-
-    @Override
-    public ExerciseDto createExercise(ExerciseDto exerciseDto, Long workoutId) {
-        User user = SecurityUtil.getPrincipalFromSecurityContext();
-        Optional<Workout> workout = Optional.ofNullable(workoutRepository.findByIdAndUser(workoutId, user)
+        Optional<Workout> workout = Optional.of(workoutRepository.findByIdAndUserId(workoutId, user.getId())
                 .orElseThrow(() -> new WorkoutNotFoundException(RequestFailure.PUT_REQUEST_FAILURE)));
 
         Exercise exercise = new Exercise();
@@ -65,18 +41,17 @@ public class ExerciseServiceImpl implements ExerciseService{
 
         Exercise newExercise = exerciseRepository.save(exercise);
 
-        ExerciseDto exerciseResponse = exerciseMapper.mapToDto(newExercise);
-
-        return exerciseResponse;
+        return exerciseMapper.mapToDto(newExercise);
     }
 
     @Override
-    public ExerciseDto updateExerciseInWorkoutById(ExerciseDto exerciseDto, Long workoutId, Long exerciseId) {
+    public ExerciseDto updateExerciseInWorkoutById(ExerciseDto exerciseDto, Long exerciseId, Long workoutId) {
         User user = SecurityUtil.getPrincipalFromSecurityContext();
-        Optional<Workout> workout = Optional.ofNullable(workoutRepository.findByIdAndUser(workoutId, user)
-                .orElseThrow(() -> new WorkoutNotFoundException(RequestFailure.PUT_REQUEST_FAILURE)));
 
-        Optional<Exercise> exercise = Optional.ofNullable(exerciseRepository.findByIdAndWorkout(exerciseId, workout.get())
+        if (!workoutRepository.existsByIdAndUserId(workoutId, user.getId())) {
+            throw new WorkoutNotFoundException(RequestFailure.GET_REQUEST_FAILURE);
+        }
+        Optional<Exercise> exercise = Optional.of(exerciseRepository.findByIdAndWorkoutId(exerciseId, workoutId)
                 .orElseThrow(() -> new ExerciseNotFoundException(RequestFailure.PUT_REQUEST_FAILURE)));
 
         Exercise updateExercise = exercise.get();
@@ -89,13 +64,15 @@ public class ExerciseServiceImpl implements ExerciseService{
     }
 
     @Override
-    public void deleteExerciseInWorkoutById(Long workoutId, Long exerciseId) {
+    public void deleteExerciseInWorkoutById(Long exerciseId, Long workoutId) {
         User user = SecurityUtil.getPrincipalFromSecurityContext();
-        Optional<Workout> workout = Optional.ofNullable(workoutRepository.findByIdAndUser(workoutId, user)
-                .orElseThrow(() -> new WorkoutNotFoundException(RequestFailure.DELETE_REQUEST_FAILURE)));
+        if (!workoutRepository.existsByIdAndUserId(workoutId, user.getId())) {
+            throw new WorkoutNotFoundException(RequestFailure.DELETE_REQUEST_FAILURE);
+        }
 
-        Optional<Exercise> exercise = Optional.ofNullable(exerciseRepository.findByIdAndWorkout(exerciseId, workout.get()))
-                .orElseThrow(() -> new ExerciseNotFoundException(RequestFailure.DELETE_REQUEST_FAILURE));
+        if (!exerciseRepository.existsByIdAndWorkoutId(exerciseId, workoutId)) {
+            throw new ExerciseNotFoundException(RequestFailure.DELETE_REQUEST_FAILURE);
+        }
         exerciseRepository.deleteById(exerciseId);
     }
